GH-92 - moved integration tests for unauthorized, if not logged in, to separate WebMvcTest - this saves test execution time because embedded mongo must not be spinned up
