+++ backend/src/test/java/com/github/chuettenrauch/mixifyapi/integration/mixtape/controller/MixtapeControllerTest.java
@@ -52,12 +52,6 @@ class MixtapeControllerTest {
     }
 
     @Test
-    void create_whenNotLoggedIn_thenReturnUnauthorized() throws Exception {
-        this.mvc.perform(post("/api/mixtapes"))
-                .andExpect(status().isUnauthorized());
-    }
-
-    @Test
     @DirtiesContext
     void create_whenLoggedIn_thenReturnMixtape() throws Exception {
         // given
@@ -131,13 +125,6 @@ class MixtapeControllerTest {
 
     @Test
     @DirtiesContext
-    void getAll_whenNotLoggedIn_thenReturnUnauthorized() throws Exception {
-        this.mvc.perform(get("/api/mixtapes"))
-                .andExpect(status().isUnauthorized());
-    }
-
-    @Test
-    @DirtiesContext
     void getAll_whenLoggedIn_thenReturnEmptyListIfNoMixtapesForTheLoggedInUserExist() throws Exception {
         // given
         OAuth2User oAuth2User = this.testUserHelper.createLoginUser();
@@ -194,12 +181,6 @@ class MixtapeControllerTest {
     }
 
     @Test
-    void delete_whenNotLoggedIn_thenReturnUnauthorized() throws Exception {
-        this.mvc.perform(delete("/api/mixtapes/123"))
-                .andExpect(status().isUnauthorized());
-    }
-
-    @Test
     void delete_whenLoggedInButMixtapeDoesNotExist_thenReturnNotFound() throws Exception {
         OAuth2User oAuth2User = this.testUserHelper.createLoginUser();
 
@@ -237,15 +218,6 @@ class MixtapeControllerTest {
                         .with(oauth2Login().oauth2User(oAuth2User))
                 )
                 .andExpect(status().isOk());
-    }
-
-    @Test
-    void update_whenNotLoggedIn_thenReturnUnauthorized() throws Exception {
-        this.mvc.perform(put("/api/mixtapes/123")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content("{}")
-                )
-                .andExpect(status().isUnauthorized());
     }
 
     @Test
@@ -348,12 +320,6 @@ class MixtapeControllerTest {
                 )
                 .andExpect(status().isOk())
                 .andExpect(content().json(expectedJson));
-    }
-
-    @Test
-    void get_whenNotLoggedIn_thenReturnUnauthorized() throws Exception {
-        this.mvc.perform(get("/api/mixtapes/123"))
-                .andExpect(status().isUnauthorized());
     }
 
     @Test
