+++ backend/src/test/java/com/github/chuettenrauch/mixifyapi/integration/mixtape/controller/TrackControllerTest.java
@@ -51,12 +51,6 @@ class TrackControllerTest {
     }
 
     @Test
-    void create_whenNotLoggedIn_thenReturnUnauthorized() throws Exception {
-        this.mvc.perform(post("/api/mixtapes/123/tracks"))
-                .andExpect(status().isUnauthorized());
-    }
-
-    @Test
     @DirtiesContext
     void create_whenMixtapeDoesNotExist_thenReturnNotFound() throws Exception {
         // given
@@ -180,12 +174,6 @@ class TrackControllerTest {
     }
 
     @Test
-    void update_whenNotLoggedIn_thenReturnUnauthorized() throws Exception {
-        this.mvc.perform(put("/api/mixtapes/123/tracks/123"))
-                .andExpect(status().isUnauthorized());
-    }
-
-    @Test
     void update_whenMixtapeNotFound_thenReturnNotFound() throws Exception {
         // given
         OAuth2User oAuth2User = this.testUserHelper.createLoginUser();
@@ -305,12 +293,6 @@ class TrackControllerTest {
                 )
                 .andExpect(status().isOk())
                 .andExpect(content().json(givenJson));
-    }
-
-    @Test
-    void delete_whenNotLoggedIn_thenReturnUnauthorized() throws Exception {
-        this.mvc.perform(delete("/api/mixtapes/123/tracks/123"))
-                .andExpect(status().isUnauthorized());
     }
 
     @Test
